Saving directly from octo
Now you can save directly from octo! When a filename is not present, `w`
will quote you for a name. When present, it will simply save the file. `W`
will always quote you for a new filename, and changes the current file to
that one. This fixes #7 (Make `w` accept a filename after it).